+++ pysurvival/cpp_extensions/survival_forest.cpp
@@ -1046,8 +1046,8 @@ size_t ForestSurvival::getTreePredictionTerminalNodeID(size_t tree_idx, size_t s
       }
 
       // Adjusting the number of threads to be <= number of cores
-      uint max_num_threads = (uint) thread::hardware_concurrency();
-      if ((num_threads < 0) | (num_threads >= max_num_threads)){
+      uint max_num_threads = thread::hardware_concurrency();
+      if ((num_threads < 0) | ((uint)num_threads >= max_num_threads)){
         num_threads = max_num_threads; 
       } else if(num_threads == 0){
             num_threads = 1; 
@@ -1157,8 +1157,8 @@ size_t ForestSurvival::getTreePredictionTerminalNodeID(size_t tree_idx, size_t s
     }
 
     // Adjusting the number of threads to be <= number of cores
-    uint max_num_threads = (uint) thread::hardware_concurrency();
-    if ((num_threads < 0) | (num_threads >= max_num_threads)){
+    uint max_num_threads = thread::hardware_concurrency();
+    if ((num_threads < 0) | ((uint)num_threads >= max_num_threads)){
       num_threads = max_num_threads; 
     } else if(num_threads == 0){
           num_threads = 1; 
@@ -1216,7 +1216,7 @@ size_t ForestSurvival::getTreePredictionTerminalNodeID(size_t tree_idx, size_t s
      std::string dependent_variable_name = this->variable_names[this->dependent_varID];
      std::string status_variable_name = this->variable_names[this->status_varID];
      std::vector<double> case_weights;
-     for (int i = 0; i < input_data.size(); ++i){
+     for (size_t i = 0; i < input_data.size(); ++i){
       case_weights.push_back(0.);
      }
      
@@ -1272,7 +1272,7 @@ size_t ForestSurvival::getTreePredictionTerminalNodeID(size_t tree_idx, size_t s
      std::string dependent_variable_name = this->variable_names[this->dependent_varID];
      std::string status_variable_name = this->variable_names[this->status_varID];
      std::vector<double> case_weights;
-     for (int i = 0; i < input_data.size(); ++i){
+     for (size_t i = 0; i < input_data.size(); ++i){
       case_weights.push_back(1./input_data.size());
      }
      
@@ -1327,7 +1327,7 @@ size_t ForestSurvival::getTreePredictionTerminalNodeID(size_t tree_idx, size_t s
      std::string dependent_variable_name = this->variable_names[this->dependent_varID];
      std::string status_variable_name = this->variable_names[this->status_varID];
      std::vector<double> case_weights;
-     for (int i = 0; i < input_data.size(); ++i){
+     for (size_t i = 0; i < input_data.size(); ++i){
       case_weights.push_back(1./input_data.size());
      }
